# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My User Project CRUD",
    "description": "My User Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "http://domain/api",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/users/auth/token": {
      "post": {
        "tags": ["users"],
        "summary": "Get auth token",
        "operationId": "Getauthtoken",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/GetauthtokenRequest"
                  },
                  {
                    "example": {
                      "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NSwiZW1haWwiOiJkZG9zLmhhY2tlckBiay5ydSIsImZpcnN0TmFtZSI6IlRlc3ROYW1lIiwibGFzdE5hbWUiOiJUZXN0U3VybmFtZSIsImlzRW1haWxWZXJpZmllZCI6dHJ1ZSwiaWF0IjoxNzAwNjE4NjkwLCJleHAiOjE3MDEyMjM0OTB9.EwITmubyudFIrCBfDqxmw5t-Wu4wiFHqe_5R1cMWWFQ"
                    }
                  }
                ]
              },
              "example": {
                "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NSwiZW1haWwiOiJkZG9zLmhhY2tlckBiay5ydSIsImZpcnN0TmFtZSI6IlRlc3ROYW1lIiwibGFzdE5hbWUiOiJUZXN0U3VybmFtZSIsImlzRW1haWxWZXJpZmllZCI6dHJ1ZSwiaWF0IjoxNzAwNjE4NjkwLCJleHAiOjE3MDEyMjM0OTB9.EwITmubyudFIrCBfDqxmw5t-Wu4wiFHqe_5R1cMWWFQ"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://example.com/:{port}/api",
            "variables": {
              "port": {
                "default": "DefaultParameterValue"
              }
            }
          }
        ]
      }
    },
...
    "/users/forgot-password": {
      "patch": {
        "tags": ["users"],
        "summary": "Forgot password",
        "operationId": "Forgotpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ForgotpasswordRequest"
                  },
                  {
                    "example": {
                      "email": "ddos.hacker@bk.r"
                    }
                  }
                ]
              },
              "example": {
                "email": "ddos.hacker@bk.r"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://example.com/:{port}/api",
            "variables": {
              "port": {
                "default": "DefaultParameterValue"
              }
            }
          }
        ]
      },
      "post": {
        "tags": ["users"],
        "summary": "Reset password",
        "operationId": "Resetpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ResetpasswordRequest"
                  },
                  {
                    "example": {
                      "email": "ddos.hacker@bk.ru"
                    }
                  }
                ]
              },
              "example": {
                "email": "ddos.hacker@bk.ru"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://example.com/:{port}/api",
            "variables": {
              "port": {
                "default": "DefaultParameterValue"
              }
            }
          }
        ]
      }
    },
    "/users/change-password/{passwordToken}": {
      "patch": {
        "tags": ["users"],
        "summary": "Change password",
        "operationId": "Changepassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ChangepasswordRequest"
                  },
                  {
                    "example": {
                      "password1": "Test-1",
                      "password2": "Test-1"
                    }
                  }
                ]
              },
              "example": {
                "password1": "Test-1",
                "password2": "Test-1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://example.com/:{port}/api",
            "variables": {
              "port": {
                "default": "DefaultParameterValue"
              }
            }
          }
        ]
      }
    },
...
    "/users/confirm-change-email/{changeEmailToken}": {
      "patch": {
        "tags": ["users"],
        "summary": "Confirm change email",
        "operationId": "Confirmchangeemail",
        "parameters": [
          {
            "name": "changeEmailToken",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://example.com/:{port}/api",
            "variables": {
              "port": {
                "default": "DefaultParameterValue"
              }
            }
          }
        ]
      }
    },
...
  },
...
}